requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==1.10.13
environs==9.5.3
python-dotenv==1.0.0
marshmallow==3.20.1